fix bug

+++ src/main.py
@@ -142,11 +142,11 @@ if __name__ == "__main__":
     if arguments.num_clusters == 1:
         n_modes = 1
     elif arguments.num_clusters == 2:
-        n_modes = len(targets) / 2
+        n_modes = int(len(targets) / 2)
     elif arguments.num_clusters == 4:
         n_modes = len(targets)
     else:
-        n_modes = len(targets) * 2
+        n_modes = int(len(targets) * 2)
     if number_of_params == 2:
         listener = VizListener(file_path=".".join([str(arguments.clustering), str(seed), str(n_modes),
                                                    str(arguments.num_dims), str(arguments.num_targets), "txt"]),
